Moved GEOM based functions to AdvancedConstruct.py
The latter requires that the GEOM wrapper is built.
